requirements.txt
Flask==2.3.3
MetaTrader5==5.0.45
python-dotenv==1.0.0
requests==2.31.0
numpy<2.0
python-dotenv